requirements.txt
aiogram
fluent.runtime
SQLAlchemy
asyncpg
psycopg2-binary
python-dotenv
redis
apscheduler
alembic
Pillow